passed tests and implemented func: - finishShift()

+++ swift/master-mixologist/Sources/MasterMixologist/MasterMixologist.swift
@@ -73,7 +73,7 @@ func makeWedges(needed: Int, limes: [String]) -> Int {
 
 func finishShift(minutesLeft: Int, remainingOrders: [[String]]) -> [[String]] {
 //  fatalError("Please implement the finishShift(minutesLeft:remainingOrders:) function")
-    var copyMinutesLeft:Int = minutesLeft
+    var copyMinutesLeft:Double = Double(minutesLeft)
     var outputRemainingOrders: [[String]] = remainingOrders
     
     repeat {
